+++ cores/rpcs/rpc_method.h
@@ -12,13 +12,13 @@ public:
     void setTxType(uchar txType) {mTxType=txType;}
 
 public slots:
-    Q_INVOKABLE double pduDataGet(uchar addr, uchar type, uchar topic, uchar sub, uchar id);
-    Q_INVOKABLE bool pduDataSet(uchar addr, uchar type, uchar topic, uchar sub, uchar id, double value);
-    Q_INVOKABLE bool pduCfgSet(uchar type, uchar fc, const QVariant &value, uchar id=0, uchar addr=0);
-    Q_INVOKABLE QString pduCfgGet(uchar type, uchar fc, uchar id=0, uchar addr=0);
-    Q_INVOKABLE QString pduLogFun(uchar type, uchar fc, int id=0, int cnt=0);
-    Q_INVOKABLE bool pduRelaysCtrl(int addr, int start, int num, uchar on);
-    Q_INVOKABLE QString pduMetaData(uchar addr=0);
+    Q_INVOKABLE double pduDataGet(int addr, int type, int topic, int sub, int id);
+    Q_INVOKABLE bool pduDataSet(int addr, int type, int topic, int sub, int id, double value);
+    Q_INVOKABLE bool pduCfgSet(int type, int fc, const QVariant &value, int id=0, int addr=0);
+    Q_INVOKABLE QString pduCfgGet(int type, int fc, int id=0, int addr=0);
+    Q_INVOKABLE QString pduLogFun(int type, int fc, int id=0, int cnt=0);
+    Q_INVOKABLE bool pduRelaysCtrl(int addr, int start, int num, int on);
+    Q_INVOKABLE QString pduMetaData(int addr=0);
     Q_INVOKABLE QString execute(const QString &cmd);
 
 signals:
